requirements.txt
allure-pytest==2.13.3
allure-python-commons==2.13.3
attrs==23.2.0
exceptiongroup==1.2.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
tomli==2.0.1
